requirements.txt
streamlit==1.32.0
opencv-python==4.9.0.80
numpy==1.26.4
mahotas==1.4.13
Pillow==10.2.0
matplotlib==3.8.3 